+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added a link to manage the benchmarks to the benchmark comparator
+- Added support for localized routes
 
 ## 1.272.0 - 2023-05-26